fix: reorganize to fix prod build
