+++ src/page-support/prob-revshare/components/active-view.js
